src/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
authlib
python-jose[cryptography]
psycopg2-binary
